Fix issue during repr of items that have defined representation
Currently there is also another bug where spines are not represented
properly? Not sure what's going on but we are almost there when it comes
to checking prelude.. My life feels like a huge dejavu. Just search
the commit history for "prelude" and see how many times I have had to
fix the typechecker to check the prelude (after each time a change is
made to the core..)